Move solving logic to puzzle solver, use same overlay
